Move all header encoding tests to a separate test file and revert original tests.
